refactor: 도메인 분리
