+++ Notechigima/src/main/java/com/capstone/notechigima/controller/GroupController.java
@@ -3,8 +3,8 @@ package com.capstone.notechigima.controller;
 import com.capstone.notechigima.config.BaseException;
 import com.capstone.notechigima.config.BaseResponse;
 import com.capstone.notechigima.config.BaseResponseStatus;
-import com.capstone.notechigima.model.dto.group.PostGroupRequestDTO;
-import com.capstone.notechigima.model.dto.users.GetUserResponseDTO;
+import com.capstone.notechigima.model.group.PostGroupRequestDTO;
+import com.capstone.notechigima.model.users.GetUserResponseDTO;
 import com.capstone.notechigima.service.GroupService;
 import com.capstone.notechigima.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
