+++ Notechigima/src/main/java/com/capstone/notechigima/repository/CommentRepository.java
@@ -1,7 +1,7 @@
 package com.capstone.notechigima.repository;
 
-import com.capstone.notechigima.model.dao.comment.CommentDetailEntity;
-import com.capstone.notechigima.model.dao.comment.CommentWriteEntity;
+import com.capstone.notechigima.domain.CommentDetailEntity;
+import com.capstone.notechigima.domain.CommentWriteEntity;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
