+++ Notechigima/src/main/java/com/capstone/notechigima/repository/NoteRepository.java
@@ -1,8 +1,8 @@
 package com.capstone.notechigima.repository;
 
-import com.capstone.notechigima.model.dao.note.NoteDetailEntity;
-import com.capstone.notechigima.model.dao.note.NoteInsertEntity;
-import com.capstone.notechigima.model.dao.note.NoteOwnerEntity;
+import com.capstone.notechigima.domain.NoteDetailEntity;
+import com.capstone.notechigima.domain.NoteInsertEntity;
+import com.capstone.notechigima.domain.NoteOwnerEntity;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
