+++ Notechigima/src/main/java/com/capstone/notechigima/repository/TopicRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.notechigima.repository;
 
-import com.capstone.notechigima.model.dao.topic.TopicEntity;
+import com.capstone.notechigima.domain.TopicEntity;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
