+++ Notechigima/src/main/java/com/capstone/notechigima/service/CommentService.java
@@ -1,7 +1,7 @@
 package com.capstone.notechigima.service;
 
 import com.capstone.notechigima.config.BaseException;
-import com.capstone.notechigima.model.dto.comment.PostCommentRequestDTO;
+import com.capstone.notechigima.model.comment.PostCommentRequestDTO;
 
 public interface CommentService {
