+++ Notechigima/src/main/java/com/capstone/notechigima/service/CommentServiceImpl.java
@@ -1,8 +1,8 @@
 package com.capstone.notechigima.service;
 
 import com.capstone.notechigima.config.BaseException;
-import com.capstone.notechigima.model.dao.comment.CommentWriteEntity;
-import com.capstone.notechigima.model.dto.comment.PostCommentRequestDTO;
+import com.capstone.notechigima.domain.CommentWriteEntity;
+import com.capstone.notechigima.model.comment.PostCommentRequestDTO;
 import com.capstone.notechigima.repository.CommentRepository;
 import org.springframework.stereotype.Service;
