+++ Notechigima/src/main/java/com/capstone/notechigima/service/GroupService.java
@@ -1,8 +1,8 @@
 package com.capstone.notechigima.service;
 
 
-import com.capstone.notechigima.model.dto.group.GetGroupResponseDTO;
-import com.capstone.notechigima.model.dto.group.PostGroupRequestDTO;
+import com.capstone.notechigima.model.group.GetGroupResponseDTO;
+import com.capstone.notechigima.model.group.PostGroupRequestDTO;
 
 import java.util.List;
