+++ Notechigima/src/main/java/com/capstone/notechigima/service/NoteServiceImpl.java
@@ -2,13 +2,13 @@ package com.capstone.notechigima.service;
 
 import com.capstone.notechigima.config.BaseException;
 import com.capstone.notechigima.model.ModelMapper;
-import com.capstone.notechigima.model.dao.note.NoteDetailEntity;
-import com.capstone.notechigima.model.dao.sentence.SentenceEntity;
-import com.capstone.notechigima.model.dao.note.NoteInsertEntity;
-import com.capstone.notechigima.model.dto.note.GetNoteResponseDTO;
-import com.capstone.notechigima.model.dto.note.GetNoteSummarizedDTO;
-import com.capstone.notechigima.model.dto.note.PostNoteRequestDTO;
-import com.capstone.notechigima.model.dto.sentence.SentenceResponseDTO;
+import com.capstone.notechigima.domain.NoteDetailEntity;
+import com.capstone.notechigima.domain.SentenceEntity;
+import com.capstone.notechigima.domain.NoteInsertEntity;
+import com.capstone.notechigima.model.note.GetNoteResponseDTO;
+import com.capstone.notechigima.model.note.GetNoteSummarizedDTO;
+import com.capstone.notechigima.model.note.PostNoteRequestDTO;
+import com.capstone.notechigima.model.sentence.SentenceResponseDTO;
 import com.capstone.notechigima.repository.NoteRepository;
 import com.capstone.notechigima.repository.SentenceRepository;
 import org.springframework.stereotype.Service;
