+++ Notechigima/src/main/java/com/capstone/notechigima/service/TopicService.java
@@ -1,6 +1,6 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.dto.topic.TopicResponseDTO;
+import com.capstone.notechigima.model.topic.TopicResponseDTO;
 import org.springframework.scheduling.annotation.Async;
 
 import java.util.List;
