+++ Notechigima/src/main/java/com/capstone/notechigima/service/UserService.java
@@ -1,6 +1,6 @@
 package com.capstone.notechigima.service;
 
-import com.capstone.notechigima.model.dto.users.GetUserResponseDTO;
+import com.capstone.notechigima.model.users.GetUserResponseDTO;
 
 import java.util.List;
