+++ Notechigima/src/main/resources/mapper/AdviceMapper.xml
@@ -13,7 +13,7 @@
         </foreach>
     </insert>
     
-    <resultMap id="adviceList" type="com.capstone.notechigima.model.dao.advice.AdviceDetailEntity">
+    <resultMap id="adviceList" type="com.capstone.notechigima.domain.adviceailEntity">
         <result property="adviceId" column="adviceId"/>
         <result property="adviceType" column="adviceType"/>
         <result property="sentenceId1" column="sentenceId1"/>
@@ -24,10 +24,10 @@
         <result property="sentence2" column="sentence2"/>
         <result property="writerId2" column="writerId2"/>
         <result property="writerName2" column="writerName2"/>
-        <collection column="adviceId" property="comments" javaType="java.util.ArrayList" ofType="com.capstone.notechigima.model.dao.comment.CommentDetailEntity" select="getCommentListById"/>
+        <collection column="adviceId" property="comments" javaType="java.util.ArrayList" ofType="com.capstone.notechigima.domain.CommentDetailEntity" select="getCommentListById"/>
     </resultMap>
 
-    <select id="getCommentListById" parameterType="java.util.Map" resultType="com.capstone.notechigima.model.dao.comment.CommentDetailEntity">
+    <select id="getCommentListById" parameterType="java.util.Map" resultType="com.capstone.notechigima.domain.CommentDetailEntity">
         SELECT c.comment_id commentId, c.user_id userId,
                u.nickname userName, c.content content,
                c.created_at createdAt, c.updated_at updatedAt
